+++ fleetpings/form.py
@@ -1,0 +1,25 @@
+# coding=utf-8
+
+"""
+form declarations
+"""
+
+from django.forms import ModelForm, TextInput
+from fleetpings.models import FleetType
+
+
+class FleetTypeAdminForm(ModelForm):
+    """
+    form definitions for the FleetType form in admin
+    """
+
+    class Meta:
+        """
+        Meta
+        """
+
+        model = FleetType
+        fields = "__all__"
+        widgets = {
+            "embed_color": TextInput(attrs={"type": "color"}),
+        }